+++ resources/views/layouts/master.blade.php
@@ -5,7 +5,7 @@
         <meta http-equiv="X-UA-Compatible" content="IE=edge">
         <meta name="viewport" content="width=device-width, initial-scale=1">
         <title>Filetinmel</title>
-        <link rel="shortcut icon" href="{{asset('favicon.ico')}}">
+        <link rel="shortcut icon" href="{{asset('images/favicon.ico')}}">
         <!-- CSRF Token -->
         <meta name="csrf-token" content="{{ csrf_token() }}">