+++ routes/api.php
@@ -1,20 +1,20 @@
 <?php
 
-use Illuminate\Http\Request;
-
 //backend routes
 if (config('filetinmel.backend')) {
 
     Route::prefix('api/filetinmel')
-        ->middleware('api:auth')
+        ->middleware('api')
         ->name('api-filetinmel.')
         ->group(function () {
 
-            Route::get('/', function (Request $request) {
-                return $request->user();
-            })->name('index');
+            Route::post('youtube', '\\Mrlinnth\\Filetinmel\\Http\\Controllers\\Api\\UploadController@postYoutube')->name('youtube');
 
-            // Route::apiResource('/plans', '\\Mrlinnth\\Filetinmel\\Http\\Controllers\\Api\\PlanController');
+            Route::post('upload/{f?}', '\\Mrlinnth\\Filetinmel\\Http\\Controllers\\Api\\UploadController@postFiles')->name('upload');
+
+            Route::post('files', '\\Mrlinnth\\Filetinmel\\Http\\Controllers\\Api\\UploadController@getFiles')->name('files');
+
+            Route::get('temp', '\\Mrlinnth\\Filetinmel\\Http\\Controllers\\Api\\UploadController@temp')->name('temp');
 
         });