+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.1.0] - 2020-04-07
+
+### Added
+- Single File component with s3 upload
+- Multiple File component with youtube upload
+
 ## [0.0.0] - 2020-03-14
 
 Project created